backend/requirements.txt
# requirements.txt - UPDATED FOR COMPATIBILITY

# ============================================================================
# SPECTRA-AI - Production Dependencies (Compatible Versions)
# ============================================================================

# Core Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Environment & Configuration
python-dotenv==1.0.0

# Database
pymongo==4.6.1
dnspython==2.6.1

# Authentication & Security
firebase-admin==6.4.0
PyJWT==2.8.0
cryptography==41.0.7

# AI & Machine Learning (UPDATED VERSIONS)
google-generativeai==0.3.2
openai-whisper==20231117
sentence-transformers==2.3.1
transformers==4.36.2

# PyTorch - Install separately based on your system
# See installation command below
# torch>=2.2.0
# torchaudio>=2.2.0

# Vector Database & Search
faiss-cpu==1.7.4
numpy>=1.24.0,<2.0.0

# YouTube & Media Processing
yt-dlp>=2024.3.10
youtube-transcript-api==0.6.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0

# HTTP & Networking
requests==2.31.0
urllib3>=2.0.0,<3.0.0

# Data Processing
scikit-learn>=1.3.0

# Logging & Monitoring
colorlog==6.8.0

# Validation & Parsing
jsonschema==4.20.0

# Date & Time
python-dateutil==2.8.2
pytz>=2023.3

# File Handling
filetype==1.2.0

# Performance & Caching
cachetools==5.3.2

# Deployment & Production
gunicorn==21.2.0

# Utilities
click>=8.1.0
tqdm>=4.66.0

# Security
bcrypt==4.1.2

# ============================================================================
# TESTING (Development Only)
# ============================================================================
# pytest==7.4.3
# pytest-flask==1.3.0

# ============================================================================
# CODE QUALITY (Development Only)
# ============================================================================
# black==23.12.1
# flake8==7.0.0


youtube-transcript-api==0.6.2
faster-whisper==1.0.3